# Optimize postgres runner for db tests