--- a/atc/db/lock/lock_suite_test.go
+++ b/atc/db/lock/lock_suite_test.go
@@ -27,7 +27,7 @@ var _ = BeforeSuite(func() {
 
 	dbProcess = ifrit.Invoke(postgresRunner)
 
-	postgresRunner.CreateTestDB()
+	postgresRunner.InitializeTestDBTemplate()
 })
 
 var _ = AfterSuite(func() {