--- a/atc/db/lock/lock_suite_test.go
+++ b/atc/db/lock/lock_suite_test.go
@@ -1,13 +1,9 @@
 package lock_test
 
 import (
-	"os"
-	"time"
-
 	"github.com/concourse/concourse/atc/postgresrunner"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	"github.com/tedsuo/ifrit"
 
 	"testing"
 )
@@ -18,19 +14,5 @@ func TestLock(t *testing.T) {
 }
 
 var postgresRunner postgresrunner.Runner
-var dbProcess ifrit.Process
 
-var _ = BeforeSuite(func() {
-	postgresRunner = postgresrunner.Runner{
-		Port: 5433 + GinkgoParallelNode(),
-	}
-
-	dbProcess = ifrit.Invoke(postgresRunner)
-
-	postgresRunner.InitializeTestDBTemplate()
-})
-
-var _ = AfterSuite(func() {
-	dbProcess.Signal(os.Interrupt)
-	Eventually(dbProcess.Wait(), 10*time.Second).Should(Receive())
-})
+var _ = postgresrunner.GinkgoRunner(&postgresRunner)