--- a/atc/db/migration/migration_suite_test.go
+++ b/atc/db/migration/migration_suite_test.go
@@ -29,7 +29,7 @@ var _ = BeforeSuite(func() {
 })
 
 var _ = BeforeEach(func() {
-	postgresRunner.CreateTestDB()
+	postgresRunner.CreateEmptyTestDB()
 })
 
 var _ = AfterEach(func() {