--- a/atc/db/migration/migration_suite_test.go
+++ b/atc/db/migration/migration_suite_test.go
@@ -1,14 +1,10 @@
 package migration_test
 
 import (
-	"os"
-	"time"
-
 	"github.com/concourse/concourse/atc/postgresrunner"
 	"github.com/gobuffalo/packr"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	"github.com/tedsuo/ifrit"
 
 	"testing"
 )
@@ -19,14 +15,8 @@ func TestMigration(t *testing.T) {
 }
 
 var postgresRunner postgresrunner.Runner
-var dbProcess ifrit.Process
 
-var _ = BeforeSuite(func() {
-	postgresRunner = postgresrunner.Runner{
-		Port: 5433 + GinkgoParallelNode(),
-	}
-	dbProcess = ifrit.Invoke(postgresRunner)
-})
+var _ = postgresrunner.GinkgoRunner(&postgresRunner)
 
 var _ = BeforeEach(func() {
 	postgresRunner.CreateEmptyTestDB()
@@ -34,11 +24,6 @@ var _ = BeforeEach(func() {
 
 var _ = AfterEach(func() {
 	postgresRunner.DropTestDB()
-})
-
-var _ = AfterSuite(func() {
-	dbProcess.Signal(os.Interrupt)
-	Eventually(dbProcess.Wait(), 10*time.Second).Should(Receive())
 })
 
 var asset = packr.NewBox("./migrations").MustBytes