--- a/cmd/concourse/concourse_test.go
+++ b/cmd/concourse/concourse_test.go
@@ -38,10 +38,8 @@ var _ = Describe("Web Command", func() {
 
 	BeforeEach(func() {
 		hostKeyFile, hostPubKeyFile, _, _ = generateSSHKeypair()
-		postgresRunner = postgresrunner.Runner{
-			Port: 5433 + GinkgoParallelNode(),
-		}
-		dbProcess = ifrit.Invoke(postgresRunner)
+		postgresrunner.InitializeRunnerForGinkgo(&postgresRunner, &dbProcess)
+
 		postgresRunner.CreateEmptyTestDB()
 
 		concourseCommand = exec.Command(
@@ -82,9 +80,7 @@ var _ = Describe("Web Command", func() {
 		<-concourseProcess.Wait()
 		postgresRunner.DropTestDB()
 
-		dbProcess.Signal(os.Interrupt)
-		err := <-dbProcess.Wait()
-		Expect(err).NotTo(HaveOccurred())
+		postgresrunner.FinalizeRunnerForGinkgo(&postgresRunner, &dbProcess)
 		os.Remove(hostKeyFile)
 		os.Remove(hostPubKeyFile)
 		os.Remove(filepath.Dir(hostPubKeyFile))