--- a/skymarshal/dexserver/dexserver_suite_test.go
+++ b/skymarshal/dexserver/dexserver_suite_test.go
@@ -1,12 +1,9 @@
 package dexserver_test
 
 import (
-	"os"
-
 	"github.com/concourse/concourse/atc/postgresrunner"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	"github.com/tedsuo/ifrit"
 
 	"testing"
 )
@@ -17,15 +14,8 @@ func TestDexServer(t *testing.T) {
 }
 
 var postgresRunner postgresrunner.Runner
-var dbProcess ifrit.Process
 
-var _ = BeforeSuite(func() {
-	postgresRunner = postgresrunner.Runner{
-		Port: 5433 + GinkgoParallelNode(),
-	}
-	dbProcess = ifrit.Invoke(postgresRunner)
-	postgresRunner.InitializeTestDBTemplate()
-})
+var _ = postgresrunner.GinkgoRunner(&postgresRunner)
 
 var _ = BeforeEach(func() {
 	postgresRunner.CreateTestDBFromTemplate()
@@ -33,9 +23,4 @@ var _ = BeforeEach(func() {
 
 var _ = AfterEach(func() {
 	postgresRunner.DropTestDB()
-})
-
-var _ = AfterSuite(func() {
-	dbProcess.Signal(os.Interrupt)
-	<-dbProcess.Wait()
 })