@aoldershaw 2021-02-01
Interestingly, it seems to make things slightly *slower* on CI - the DB suite took `3m42s` on this PR, whereas on previous PRs it took ~`3m10s`.

I wonder why it's so much faster on my laptop with these changes. Could be a result of some optimizations in `DROP DATABASE` or `CREATE DATABASE ... TEMPLATE ...` between Postgres versions (my laptop is running Postgres 13, while the unit image is using Postgres 11.