@aoldershaw 2021-02-01
> I wonder why it's so much faster on my laptop with these changes

I didn't realize we use a tmpfs on CI for postgres - it's possible that with a tmpfs the truncates are actually pretty fast, whereas on a SSD it's faster to DROP+CREATE.

It could also have been a fluke that it was slower - I hijacked into the container and reran the DB suite and it ran in just `2m54s` - not much faster, but at least not slower. I also tried bumping postgres to 13, and it ran in `2m29s` - so maybe we should bump postgres in the unit image (I don't think there's any particular reason to test against Postgres 11 in unit)