@aoldershaw 2021-02-01
I upgraded the Postgres version to 13 on the unit image, and the DB suite ran in `2m9s`, about 1 minute faster than the typical duration before bumping Postgres and these optimizations. I ran another PR with Postgres 13 (without the optimizations), and it's still around 3 minutes, so there's at least some improvement on CI!